.github/workflows/requirements.txt
affine==2.4.0
albucore==0.0.23
albumentations==2.0.2
annotated-types==0.7.0
attrs==25.1.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
click-plugins==1.1.1
cligj==0.7.2
contourpy==1.3.1
cycler==0.12.1
filelock==3.13.1
fonttools==4.55.8
fsspec==2024.6.1
huggingface-hub==0.28.0
idna==3.10
Jinja2==3.1.4
kiwisolver==1.4.8
MarkupSafe==2.1.5
matplotlib==3.10.0
mpmath==1.3.0
networkx==3.3
numpy==2.1.2
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.21.5
nvidia-nvtx-cu11==11.8.86
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==24.2
pillow==11.0.0
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
rasterio==1.4.3
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
scipy==1.15.1
simsimd==6.2.1
six==1.17.0
stringzilla==3.11.3
sympy==1.13.1
tokenizers==0.21.0
torch==2.6.0+cu118
torchaudio==2.6.0+cu118
torchvision==0.21.0+cu118
tqdm==4.67.1
transformers==4.48.1
triton==3.2.0
typing_extensions==4.12.2
urllib3==2.3.0
